+++ library/core/src/main/java/com/google/android/exoplayer2/text/ssa/SsaStyle.java
@@ -95,6 +95,8 @@ import java.util.regex.Pattern;
   public final float fontSize;
   public final boolean bold;
   public final boolean italic;
+  public final boolean underline;
+  public final boolean strikeout;
 
   private SsaStyle(
       String name,
@@ -102,13 +104,17 @@ import java.util.regex.Pattern;
       @Nullable @ColorInt Integer primaryColor,
       float fontSize,
       boolean bold,
-      boolean italic) {
+      boolean italic,
+      boolean underline,
+      boolean strikeout) {
     this.name = name;
     this.alignment = alignment;
     this.primaryColor = primaryColor;
     this.fontSize = fontSize;
     this.bold = bold;
     this.italic = italic;
+    this.underline = underline;
+    this.strikeout = strikeout;
   }
 
   @Nullable
@@ -136,10 +142,16 @@ import java.util.regex.Pattern;
               ? parseFontSize(styleValues[format.fontSizeIndex].trim())
               : Cue.DIMEN_UNSET,
           format.boldIndex != C.INDEX_UNSET
-              ? parseBoldOrItalic(styleValues[format.boldIndex].trim())
+              ? parseBooleanValue(styleValues[format.boldIndex].trim())
               : false,
           format.italicIndex != C.INDEX_UNSET
-              ? parseBoldOrItalic(styleValues[format.italicIndex].trim())
+              ? parseBooleanValue(styleValues[format.italicIndex].trim())
+              : false,
+          format.underlineIndex != C.INDEX_UNSET
+              ? parseBooleanValue(styleValues[format.underlineIndex].trim())
+              : false,
+          format.strikeoutIndex != C.INDEX_UNSET
+              ? parseBooleanValue(styleValues[format.strikeoutIndex].trim())
               : false);
     } catch (RuntimeException e) {
       Log.w(TAG, "Skipping malformed 'Style:' line: '" + styleLine + "'", e);
@@ -226,12 +238,12 @@ import java.util.regex.Pattern;
     }
   }
 
-  private static boolean parseBoldOrItalic(String boldOrItalic) {
+  private static boolean parseBooleanValue(String booleanValue) {
     try {
-      int value = Integer.parseInt(boldOrItalic);
+      int value = Integer.parseInt(booleanValue);
       return value == 1 || value == -1;
     } catch (NumberFormatException e) {
-      Log.w(TAG, "Failed to parse bold/italic: '" + boldOrItalic + "'", e);
+      Log.w(TAG, "Failed to parse boolean value: '" + booleanValue + "'", e);
       return false;
     }
   }
@@ -250,6 +262,8 @@ import java.util.regex.Pattern;
     public final int fontSizeIndex;
     public final int boldIndex;
     public final int italicIndex;
+    public final int underlineIndex;
+    public final int strikeoutIndex;
     public final int length;
 
     private Format(
@@ -259,6 +273,8 @@ import java.util.regex.Pattern;
         int fontSizeIndex,
         int boldIndex,
         int italicIndex,
+        int underlineIndex,
+        int strikeoutIndex,
         int length) {
       this.nameIndex = nameIndex;
       this.alignmentIndex = alignmentIndex;
@@ -266,6 +282,8 @@ import java.util.regex.Pattern;
       this.fontSizeIndex = fontSizeIndex;
       this.boldIndex = boldIndex;
       this.italicIndex = italicIndex;
+      this.underlineIndex = underlineIndex;
+      this.strikeoutIndex = strikeoutIndex;
       this.length = length;
     }
 
@@ -282,6 +300,8 @@ import java.util.regex.Pattern;
       int fontSizeIndex = C.INDEX_UNSET;
       int boldIndex = C.INDEX_UNSET;
       int italicIndex = C.INDEX_UNSET;
+      int underlineIndex = C.INDEX_UNSET;
+      int strikeoutIndex = C.INDEX_UNSET;
       String[] keys =
           TextUtils.split(styleFormatLine.substring(SsaDecoder.FORMAT_LINE_PREFIX.length()), ",");
       for (int i = 0; i < keys.length; i++) {
@@ -304,6 +324,12 @@ import java.util.regex.Pattern;
           case "italic":
             italicIndex = i;
             break;
+          case "underline":
+            underlineIndex = i;
+            break;
+          case "strikeout":
+            strikeoutIndex = i;
+            break;
         }
       }
       return nameIndex != C.INDEX_UNSET
@@ -314,6 +340,8 @@ import java.util.regex.Pattern;
               fontSizeIndex,
               boldIndex,
               italicIndex,
+              underlineIndex,
+              strikeoutIndex,
               keys.length)
           : null;
     }